chore(test): make use of workbench service provider to customize default configs
